requirements.txt
# Document Processor Dependencies

# Core dependencies
PyMuPDF>=1.23.0  # PDF processing (fitz)
python-docx>=1.1.0  # DOCX processing

# Embedding models
sentence-transformers>=2.2.0  # Local embedding generation

# Vector store
numpy>=1.24.0  # Required for FAISS and vector operations
faiss-cpu>=1.7.4  # Fast similarity search (optional)

# API
fastapi>=0.104.0  # Web framework
uvicorn[standard]>=0.24.0  # ASGI server
python-multipart>=0.0.6  # Form data parsing
pydantic>=2.0.0  # Data validation

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: GPU support for faster embeddings
# torch>=2.0.0  # Install separately with CUDA if needed
